Add files via upload
Last Version
Big Update of iCGS
Add Case Studies, Class Back Front and Main Sepratation, 
Add Model Cheking and Interactive Graph Design
