+++ example_strategy.txt
@@ -1,0 +1,17 @@
+Transition
+0 AC,AD BC,BD 0
+0 0 AD,BD AC,BC
+0 0 0 AD,BC
+0 0 0 *
+Unkown_Transition_by
+0 0 0 0
+0 0 0 0
+0 0 0 0
+0 0 0 0
+Name_State
+s0 s1 s2 s3
+Initial_State
+s0
+
+
+
