+++ finite_state_machine.gv
@@ -1,13 +1,13 @@
 digraph finite_state_machine {
 	node [shape=circle]
-	s2 [fillcolor=red style=filled]
+	s3 [fillcolor=red style=filled]
 	sI -> sI [label="*"]
-	sI -> s1 [label="l&s"]
-	sI -> s2 [label="l&l"]
-	sI -> s3 [label="r&l"]
-	sI -> s4 [label="r&r"]
-	sI -> s5 [label="l&r"]
-	sI -> s6 [label="r&s"]
+	sI -> s1 [label=ls1]
+	sI -> s2 [label=ll2]
+	sI -> s3 [label=rl3]
+	sI -> s4 [label=rs4]
+	sI -> s5 [label=lr5]
+	sI -> s6 [label=rr6]
 	s1 -> sI [label=a]
 	s1 -> s7 [label=o]
 	s2 -> s7 [label=o]
@@ -16,7 +16,7 @@ digraph finite_state_machine {
 	s4 -> sI [label=a]
 	s4 -> s7 [label=o]
 	s5 -> sI [label=a]
-	s5 -> s5 [label=e]
+	s5 -> s4 [label=e]
 	s5 -> s7 [label=o]
 	s6 -> s3 [label=e]
 	s6 -> s7 [label=o]
